add transcopyright
The transcopyright license is a "deliberative license". transcopyright is used for the xanadu project by Ted Nelson (one of the creators of hypertext, hypermedia etc).

The deliberative license model is different from the restrictive model (AGPL, GPLv2, GPLv3 etc) or the permissive model (MIT, BSD etc).

In the deliberative license model it is something between copyright and copyleft, something similar to what I believe is CC (Creative Commons).

For example, if the work is licensed under transcopyright and is MIT, this could include both proprietary and non-proprietary things.

If the work is licensed under transcopyright and is GPLv3, the mixed licenses include the parts that make up the GPLv3 and contrary licenses such as MIT or proprietary as well.

These two cases occur, because the transcopyright license understands that the code is composed of isolated or separate parts (from what I initially understood). So, you only license what can or should be licensed. It's a way of having a copyright in a copyleft way. I think it would be something between a "non-copyright" or "non-copyleft" - it's like a "copy-left-right-center" - this license allows mixing software, whether they have this license or another more permissive one or not.